pyproject.toml
[tool.poetry]
name = "switter"
version = "0.1.0"
description = ""
authors = ["Cheburator <sashadanilenko06@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
psycopg2 = "^2.9.9"
pillow = "^10.2.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
